+++ x/gov/types/genesis.go
@@ -1,5 +1,10 @@
 package types
 
+// special messages managed by governance
+const (
+	UpsertTokenAlias = "upsert-token-alias"
+)
+
 // DefaultGenesis returns the default CustomGo genesis state
 func DefaultGenesis() *GenesisState {
 	return &GenesisState{
@@ -9,13 +14,13 @@ func DefaultGenesis() *GenesisState {
 		},
 		StartingProposalId: 1,
 		NetworkProperties: &NetworkProperties{
-			MinTxFee: 1,
-			MaxTxFee: 10000,
+			MinTxFee: 100,
+			MaxTxFee: 1000000,
 		},
 		ExecutionFees: []*ExecutionFee{
 			{
 				Name:              "Claim Validator Seat",
-				TransactionType:   "A",
+				TransactionType:   "claim-validator-seat",
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,
@@ -23,7 +28,7 @@ func DefaultGenesis() *GenesisState {
 			},
 			{
 				Name:              "Claim Governance Seat",
-				TransactionType:   "A",
+				TransactionType:   "claim-governance-seat",
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,
@@ -31,7 +36,7 @@ func DefaultGenesis() *GenesisState {
 			},
 			{
 				Name:              "Claim Proposal Type X",
-				TransactionType:   "B",
+				TransactionType:   "claim-proposal-type-x",
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,
@@ -39,7 +44,7 @@ func DefaultGenesis() *GenesisState {
 			},
 			{
 				Name:              "Vote Proposal Type X",
-				TransactionType:   "B",
+				TransactionType:   "vote-proposal-type-x",
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,
@@ -47,7 +52,7 @@ func DefaultGenesis() *GenesisState {
 			},
 			{
 				Name:              "Submit Proposal Type X",
-				TransactionType:   "B",
+				TransactionType:   "submit-proposal-type-x",
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,
@@ -55,7 +60,15 @@ func DefaultGenesis() *GenesisState {
 			},
 			{
 				Name:              "Veto Proposal Type X",
-				TransactionType:   "B",
+				TransactionType:   "veto-proposal-type-x",
+				ExecutionFee:      10,
+				FailureFee:        1,
+				Timeout:           10,
+				DefaultParameters: 0,
+			},
+			{
+				Name:              "Upsert Token Alias Execution Fee",
+				TransactionType:   UpsertTokenAlias,
 				ExecutionFee:      10,
 				FailureFee:        1,
 				Timeout:           10,