+++ x/tokens/types/errors.go
@@ -1,0 +1,8 @@
+package types
+
+import "github.com/cosmos/cosmos-sdk/types/errors"
+
+// tokens module errors
+var (
+	ErrTokenAliasNotFound = errors.Register(ModuleName, 2, "token alias not found")
+)