+++ x/tokens/types/keys.go
@@ -1,0 +1,6 @@
+package types
+
+const ModuleName = "tokens"
+
+// RouterKey to be used for routing msgs
+const RouterKey = ModuleName